Cold temperature bug, PIC32MX564F128H
New:
- The MCU supported now is PIC32MX564F128H (was PIC32MX534F064H)

Bugfixes:
- Temperature goes to 511C and controller refuses to heat up when tip is cold.

+++ software/front/US_Firmware.X/sensorMath.h
@@ -110,7 +110,7 @@ typedef struct {
 #define SENSORMATH_H_EXTERN
 #endif
 
-INT32 GetSensorTemperature(UINT16 input, t_SensorConfig * SC);
+INT32 GetSensorTemperature(int input, t_SensorConfig * SC);
 
 #undef SENSORMATH_H_EXTERN
